% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bard_functions.R
\name{bard_query}
\alias{bard_query}
\title{Execute a BARD query}
\usage{
bard_query(
  tagIDs,
  dateStart = "2012-01-01",
  dateEnd = Sys.Date(),
  baseurl = "http://sandbox5.metro.ucdavis.edu/memo/getTagCSV/",
  curlH = curl,
  ...
)
}
\arguments{
\item{dateStart}{start date of query}

\item{dateEnd}{end date of query, defaults to today}

\item{baseurl}{BARD query homepage}

\item{curlH}{curl handle; defaults to curl, which if you've run setup.R is in the global environment; running the following code will recreate it: \code{curl = getCurlHandle(useragent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.100 Safari/537.36", followlocation = TRUE)}}

\item{tagids}{a vector of tag IDs formatted like those returned from `bard_tags()`}
}
\value{
a data frame of BARD detections; note that these detections are in a summarized format that is NOT the same as our detections.
}
\description{
Execute a BARD query
}
\examples{
bard_query(tagIDS = bard_tags(c(2841, 2842)), curlH = getCurlHandle(useragent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.100 Safari/537.36", followlocation = TRUE) )

}
