% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bard_functions.R
\name{bard_tags}
\alias{bard_tags}
\title{format TagIDs for BARD query}
\usage{
bard_tags(tagids, df, tagVar = "TagID", codeVar = "CodeSpace")
}
\arguments{
\item{tagids}{a vector of numeric TagIDs}

\item{df}{a data frame that contains (at minimum) columns for TagID and Codespace}

\item{tagVar}{The name of the TagID column in df}

\item{codeVar}{The name of the CodeSpace column in df}
}
\value{
A vector of tagids formatted for a query of the BARD database
}
\description{
format TagIDs for BARD query
}
\details{
This function assumes your tags are operating on the 69khz frequency.
}
