% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yb_residence.R
\name{calc_yb_residence}
\alias{calc_yb_residence}
\title{Calculate total residence time within the Yolo Bypass receiver array for a single detection year}
\usage{
calc_yb_residence(dets_df, TagID_col = "TagID")
}
\arguments{
\item{dets_df}{detections data frame that is the result of the tag_tales function, with a ReleaseTime column, an arrival column, and a departure column}

\item{TagID_col}{name of the tagid column}
}
\value{
A data frame of TagIDs and residence times
}
\description{
Assumes: you have a detections dataframe for a single detection year that has the following columns:
}
\details{
-  A "ReleaseTime" column that indicates the release time for each TagID, formatted as Y-m-d hh:mm:ss.

- arrival and departure columns, like those resulting from the tag_tales() function.

this function is designed to work on one detection year at at time
}
