% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vrls.R
\name{check_vrls}
\alias{check_vrls}
\title{Check VRL files}
\usage{
check_vrls(
  vrl_file_dir,
  db_filepath,
  date_range = c("2012-01-01", "2060-01-01"),
  con = dbConnect(SQLite(), db_filepath),
  vrl_files = list.files(vrl_file_dir, pattern = "\\\\.vrl$", recursive = TRUE,
    full.names = TRUE),
  serial_date_df = grab_serial_date(vrl_files),
  deploy_query = "SELECT Receiver, VRLDate FROM deployments"
)
}
\arguments{
\item{vrl_file_dir}{the parent directory for the VRL files}

\item{db_filepath}{the full file path to the database}

\item{date_range}{a vector of two quoted dates in "YYYY-MM-DD" format}

\item{con}{a connection to a SQL database}

\item{vrl_files}{a character vector of the VRL files to check
against}

\item{serial_date_df}{a data.frame created by
\code{grab_serial_date}}

\item{deploy_query}{SQL query to grab the deployment records of
interest}
}
\value{
list, with two data.frame, the original serial_date_df
    with an additional logical column specifying if the VRL record
    was present in the database, and a data.frame of the
    detections, with an additional logical column indicating if
    the record in the database has an associated file in the
    folder.
}
\description{
This checks the serial number and date of VRL records against
those present in the database.
}
\details{
This checks the serial number and date of VRL records against
those present in the database.
}
\author{
Matt Espe
}
