% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{connectGDB}
\alias{connectGDB}
\title{Connect to GDB}
\usage{
connectGDB(db_path, driver)
}
\arguments{
\item{db_path}{the full path to the database. Can be searched for
using \code{find_gdb()}}

\item{driver}{the result of RSQLite::SQLite()}
}
\value{
a connection to the specified database
}
\description{
Connect to GDB
}
\details{
This is a convienence wrapper which first checks if the database
exists before attempting to open it. This avoids sqlite creating a
new database if the path is misspecified.
}
\author{
Matt Espe
}
