% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_vrls.R
\name{copy_vrls}
\alias{copy_vrls}
\title{Copy VRLS files to new directory}
\usage{
copy_vrls(
  dir = ".",
  dest,
  file_regex = "RLD",
  file_type = "vrl",
  regex_exclude = TRUE,
  all_files = list.files(dir, pattern = paste0("\\\\.", file_type, "$"), full.names =
    TRUE, recursive = TRUE),
  match_files = grep(all_files, pattern = file_regex, invert = regex_exclude, value =
    TRUE),
  ...
)
}
\arguments{
\item{dir}{string, the base directory to begin looking for files}

\item{dest}{string, a destination directory. Will be created if it
doesn't exist}

\item{file_regex}{regex for matching}

\item{file_type}{the file extension for files to move}

\item{regex_exclude}{logical, whether the regex should match files
to exclude or include}

\item{all_files}{all prospective files to copy}

\item{match_files}{files which match the parameters from the regex
and file type}

\item{...}{additional args passed to \code{file.copy}}
}
\description{
Copies vrl files from field directory
}
\details{
Copies files
}
\author{
Matt Espe
}
