% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_init}
\alias{db_init}
\title{Initialize a database}
\usage{
db_init(
  db_file,
  driver = SQLite(),
  con = dbConnect(driver, db_file),
  schema = strsplit(paste(readLines(schema_file), collapse = " "), ";")[[1]],
  schema_file = system.file("create_db.sql", package = "ybt")
)
}
\arguments{
\item{db_file}{string, the full path and name of the database file
to be initialized.}

\item{driver}{A SQLite driver}

\item{con}{A connection to an empty database. If the database does
not exist, it will be created.}

\item{schema}{string, a SQLite command to create the database}

\item{schema_file}{a file with SQLite code specifying the schema
for the database.}
}
\description{
Initialize an empty database
}
\details{
This function initialized a database according to a schema. This
is done so that contraints on the data fields, including keys, can
be specified. Once initialized, the database can be populated
while ensuring that the data meet the constraints.
}
\author{
Matt Espe
}
