% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ultis.R
\name{find_gdb_dir}
\alias{find_gdb_dir}
\title{Find GDB directory}
\usage{
find_gdb_dir(dir_name, base_dir = "~")
}
\arguments{
\item{dir_name}{the name of the directory}

\item{base_dir}{the directory to begin the search}
}
\value{
path to directory
}
\description{
Finds the GDB directory
}
\details{
Convienence wrapper for dir()
}
\author{
Matt Espe
}
