% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VUE_csv.R
\name{format_detections}
\alias{format_detections}
\title{Format detections that have been queried from database}
\usage{
format_detections(
  dets_df,
  tagid_col = "TagID",
  rec_col = "Receiver",
  datetime_col = "DateTimeUTC"
)
}
\arguments{
\item{dets_df}{Detections dataframe as it is returned by dbGetQuery("select * from detections") in the Yolo Telemetry Project .sqlite database.}

\item{tagid_col}{Quoted name of the column with Tag or Fish IDs in it (defaults to "TagID"); this is passed to parse_tagid_col().}

\item{rec_col}{Quoted name of the column with the Receiver ID in it (defaults to "Receiver"); this is passed to parse_receiver_col().}

\item{datetime_col}{Quoted name of the column with detection date time stamps; defaults to "DateTimeUTC".}
}
\value{
A dataframe
}
\description{
This function parses the TagID and Receiver columns from the detections table of the database, formats the UTC timestamp column to "%Y-%m-%d %H:%M:%S", and creates a DateTimePST column with a Pacific/Pitcairn time zone (PST).
}
\author{
Myfanwy Johnston
}
