% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_det_window.R
\name{get_det_window}
\alias{get_det_window}
\title{Get detection window
get the range of detections for each detection year of a dataframe}
\usage{
get_det_window(dets_df, timecol)
}
\arguments{
\item{dets_df}{Detections dataframe with a Detyear column.  To add the Detyear column, call \code{get_det_year()} on your dataframe first.}

\item{timecol}{The name of the date-time column for the detections, in quotes}
}
\value{
A dataframe with the date and time of the first and last detection for each Detection Year present in the dataframe (between 2011 and 2018).
}
\description{
Get detection window
get the range of detections for each detection year of a dataframe
}
\examples{
get_det_window(chn_detections, "DateTimePST")

}
