% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_det_year.R
\name{get_det_year}
\alias{get_det_year}
\title{Get Detection Year
Add a detection year column to a detections dataframe}
\usage{
get_det_year(detsdf, timecol)
}
\arguments{
\item{detsdf}{Dataframe of detections data}

\item{timecol}{The name of the Date-time column, in quotes (like "DateTimePST")}
}
\value{
The dataframe, with a Detyear column added
}
\description{
Get Detection Year
Add a detection year column to a detections dataframe
}
\details{
Categorizes detections based on whether they fall between July 1st and June 30th of a given year.  Currently supports years 2011 - 2018; future versions of this function will be more flexible and general.
}
\examples{
get_det_year(dets, "DateTimePST")

}
