% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvenienceFxns.R
\name{get_dt}
\alias{get_dt}
\title{Get DateTagged}
\usage{
get_dt(tagid)
}
\arguments{
\item{tagid}{The (numeric) TagID of the fish for which you want the date of tagging}
}
\value{
The DateTagged value for the TagID
}
\description{
Get DateTagged
}
\details{
Note that this function requires the alltags global variable
}
\examples{
get_dt(2841)
}
