% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dup_deps.R
\name{get_dup_deps}
\alias{get_dup_deps}
\title{Get duplicate deployments - Find receivers with associated with more than one station in the deployments table}
\usage{
get_dup_deps(dep_df)
}
\arguments{
\item{dep_df}{Deployments dataframe, like the one returned from `dplyr::tbl(conn, "deployments") %>% collect()`}
}
\value{
a vector of receiver serial numbers that are associated with more than one station
}
\description{
Get duplicate deployments - Find receivers with associated with more than one station in the deployments table
}
\details{
# as of July 2019, there are only two receivers that have been used and then re-deployed in another place: Above Ag4/Belwo Wallace Weir, and Above Wallace Weir/Below Los Rios Check Dam.
}
\examples{
\dontrun{
deps <- dplyr::tbl(conn, "deployments") \%>\% collect() # where conn = yb database connection
get_dup_deps(deps)
}
}
