% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bard_functions.R
\name{get_last_bard}
\alias{get_last_bard}
\title{get final location and detection time for every tagID in the output of a bard query}
\usage{
get_last_bard(bdf, tagcol = "TagID")
}
\arguments{
\item{bdf}{Dataframe returned by a BARD query, i.e. the return object of `bard_query`. Note - `get_last_bd()` calls `parse_tagid_col()` internally, so don't call that function before you run this one.}

\item{tagcol}{Name of the TagID column in `bdf`}
}
\value{
Dataframe with columns: `TagID`, `last_det_bard` = last detection recorded for that TagID in the provided BARD dataframe, `last_loc_bard` = name of the Station associated with the last detection
}
\description{
get final location and detection time for every tagID in the output of a bard query
}
\details{
Note - `get_last_bd` calls `parse_tagid_col()` internally, so don't call that function before you run this one.
}
