% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vrls.R
\name{grab_serial_date}
\alias{grab_serial_date}
\title{Grab serial number and date}
\usage{
grab_serial_date(x)
}
\arguments{
\item{x}{a character vector of filenames of VRL files}
}
\value{
data.frame with the serial number, date, and file name
}
\description{
Grabs the serial number and date from a VRL file
}
\details{
Grabs the serial number and date from a VRL file
}
\author{
Matt Espe
}
