% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheds.R
\name{id_one_shed}
\alias{id_one_shed}
\title{Identify shed for one tag}
\usage{
id_one_shed(
  df,
  max_last_residence = 7,
  max_gap_days = 60,
  id_col = "TagID",
  stn_col = "GroupedStn",
  arrival_col = "arrival",
  depart_col = "departure"
)
}
\arguments{
\item{df}{a data.frame of detection history for a single tagID,
i.e., a single fish}

\item{max_last_residence}{integer, the max residence time allowed,
in days}

\item{id_col}{character, the name of the TagID column}

\item{stn_col}{character, the name of the stations column}

\item{arrival_col}{character, the name of the arrivals column}

\item{depart_col}{character, the name of the departures column}

\item{max_gap}{integer, the max gap allowed, in days}
}
\value{
a data.frame for a single fish, with logical columns
    indicating if this tag was likely a shed, or if it had large
    gap in the detection history
}
\description{
Identify a shed or gap in a single fish detection history.
}
\details{
Identify a shed or gap in a single fish detection history.
}
\author{
Matt Espe
}
