% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheds.R
\name{identify_sheds}
\alias{identify_sheds}
\title{Identify Shed tags}
\usage{
identify_sheds(df, id_column = "TagID", ...)
}
\arguments{
\item{df}{data.frame of detection histories.}

\item{id_column}{character, the name of the TagID column}

\item{...}{additional args passed to id_one_shed}
}
\value{
data.frame, with TagID, shed (logical), and long_gap
    (logical)
}
\description{
Identify a shed tag
}
\details{
.. content for \details{} ..
}
\author{
Matt Espe
}
