% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_and_max_stn.R
\name{min_and_max_stn}
\alias{min_and_max_stn}
\title{Get Minimum and Maximum Detection of a Tagged Fish}
\usage{
min_and_max_stn(df)
}
\arguments{
\item{df}{A detections dataframe - currently requires the following columns:
\item{TagID column called "TagID"}
\item{Date-time column called "DateTimePST"}
\item{Grouped station column called "GroupedStn}}
}
\value{
A dataframe with the TagIDs, first and last detections, and first and last stations for each fish.
}
\description{
Get Minimum and Maximum Detection of a Tagged Fish
}
\examples{
min_and_max_stn(dets)
}
