% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_receiver_col.R
\name{parse_receiver_col}
\alias{parse_receiver_col}
\title{Parse default-formatted Vemco Receiver column into two separate columns}
\usage{
parse_receiver_col(df, reccol = "Receiver")
}
\arguments{
\item{df}{dataframe with a `Receiver` column that is formatted as 'VRX-ReceiverSN' (i.e.,  must be delimited by hyphens)}

\item{tagcol}{the name of the column you wish to parse}
}
\value{
A dataframe with the `Receiver` column at the end, with the `VRX` info and hyphen removed.
}
\description{
separates the default Vemco column formatting of VR2W-RecSN, then just keeps the ReceiverSN as a new column
}
\details{
This function will not work on a column delimited by more than one hyphen
}
\examples{
df <- structure(list(TagID = c("A69-1206-112", "A69-1206-152", "A69-1206-1908"
), Monitor = c("VR2W-106683", "VR2W-106683", "VR2W-106683"),
    Detections = c(1L, 1L, 1L)), row.names = c(NA, 3L), class = "data.frame")
parse_receiver_col(df, "Monitor")
}
