% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvenienceFxns.R
\name{plot_track}
\alias{plot_track}
\title{Plot track}
\usage{
plot_track(df, tagid)
}
\arguments{
\item{df}{Dataframe of detections}

\item{tagid}{Numeric TagID of the fish whose track you want to plot}
}
\value{
A ggplot
}
\description{
A rudimentary ggplot of a fish track - depends on DateTimePST, GroupedStn, and rkms columns
}
