% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheds.R
\name{truncate_sheds}
\alias{truncate_sheds}
\title{Truncate Detection Histories}
\usage{
truncate_sheds(
  raw_detection_df,
  TagIDs,
  keep_last = 2L,
  units = "days",
  time_stamps = NULL,
  TagID_col = "TagID",
  Station_col = "GroupedStn",
  DateTime_col = "DateTimePST"
)
}
\arguments{
\item{raw_detection_df}{data.frame of raw detection histories}

\item{TagIDs}{vector, the TagIDs to truncate}

\item{keep_last}{integer, amount of detections to keep in the
final station history}

\item{units}{character, either days or hours}

\item{time_stamps}{Optional, if provided these are used as the
cutoff for each tagID. There must be one time_stamp for each
tagID.}

\item{TagID_col}{character, the column name for the TagID}

\item{Station_col}{character, the column name for the station}

\item{DateTime_col}{character, the column name for DateTime}
}
\value{
data.frame, with shed tags truncated
}
\description{
Truncates the detection histories to either a number of days or
hours after the fish first visit a station, or if a timestamp is
provided, to that timestamp.
}
\author{
Matt Espe
}
