% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ybt-package.R
\docType{package}
\name{ybt-package}
\alias{ybt-package}
\alias{_PACKAGE}
\alias{ybt}
\title{The Yolo Bypass Telemetry package (ybt)}
\description{
Database, formatting, and analysis functions for DWR's Yolo Telemetry Project.  #'    Data covers tagged fish detections from 2011-2020.
}
\keyword{internal}
