% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append.R
\name{ybt_db_append}
\alias{ybt_db_append}
\title{YBT Append Detection Observations}
\usage{
ybt_db_append(new_df, db_table, db_path, driver = SQLite())
}
\arguments{
\item{new_df}{data.frame containing the new observations to append
to the database}

\item{db_table}{character, the name of the database table to
append into}

\item{db_path}{character, the file path to the database}

\item{driver}{the SQLite driver to use}
}
\value{
NULL. This function is called for its side effects.
}
\description{
Appends new data to an existing database.
}
\details{
Appends new data to an existing database.
}
\author{
Matt Espe
}
