% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dup_deps.R
\name{get_dup_deps}
\alias{get_dup_deps}
\title{Get duplicate deployments - Find receivers with associated with more than one station in the deployments table}
\usage{
get_dup_deps(dep_df)
}
\arguments{
\item{dep_df}{Deployments dataframe, like the one returned from `dbGetQuery(con, "select * from deployments)`}
}
\value{
a vector of receiver serial numbers that are associated with more than one station
}
\description{
Get duplicate deployments - Find receivers with associated with more than one station in the deployments table
}
\details{
as of July 2019, there are only two receivers that have been used and then re-deployed in another place: Above Ag4/Below Wallace Weir, and Above Wallace Weir/Below Los Rios Check Dam.
}
\examples{
\dontrun{
deps <- dbGetQuery(con, "select * from deployments")  
get_dup_deps(deps)
}
}
