% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheds.R
\name{identify_sheds}
\alias{identify_sheds}
\title{Identify Shed tags}
\usage{
identify_sheds(df, id_column = "TagID", ...)
}
\arguments{
\item{df}{data.frame of detection histories created by \code{tag_tales()}.}

\item{id_column}{character, the name of the TagID column}

\item{...}{additional args passed to id_one_shed}
}
\value{
data.frame, with TagID, shed (logical), and long_gap
    (logical)
}
\description{
Identify a shed tag
}
\details{
This function identifies a shed tag, or a large gap in a fish
history, according to if a tag was detected at a single station
for 7 continuous days, without a break in the detections for more
than one hour.

The \code{tag_tales()} function defaults to lumping all detections
with no more than 1 hour between detections into a single stay. By
default, the shed tag can only be at the last detected station (as
the tag should not be moving once shed).
}
\author{
Matt Espe
}
