% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append.R
\name{ybt_db_append}
\alias{ybt_db_append}
\title{YBT Append Detection Observations}
\usage{
ybt_db_append(new_df, db_table, db_path, driver = SQLite())
}
\arguments{
\item{new_df}{}

\item{db_table}{}

\item{db_path}{}

\item{driver}{}
}
\value{

}
\description{
Appends new data to an existing database.
}
\details{
Appends new data to an existing database.
}
\author{
Matt Espe
}
